+++ django_spire/seeding/__init__.py
@@ -1,0 +1,3 @@
+from django_spire.seeding.model.django.seeder import DjangoModelSeeder
+
+__all__ = ["DjangoModelSeeder"]