+++ example/settings.py
@@ -110,6 +110,7 @@ MIDDLEWARE = [
 
 DEFAULT_AUTO_FIELD = 'django.db.models.AutoField'
 
+
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.postgresql',